Добавляет модули (#15)
* Исправляет функции

* удаляет вывод в консоль

* Исправляет функию2

* Добавляет модули
